refactor(backup): split json module into encryption/key/model/format
Break the monolithic json.rs into focused modules: encryption (seal/open
envelope), key (BackupKey derivation), model (backup data types), and
format/json (JSON framing). No behavior change; full backup suite green.
